Octree-like AMR support part 1
Added a radmc3dTree class to handle octree-like grids with arbitrary
number of active dimensions.

I changed the grid initialisation in the radmc3dData class. Instead of
an empty radmc3dGrid instance now the radmc3dData.grid will be
initialised with None. Thus the grid type will be decided upon
reading or creating the grid.

Added the octree option to the formatted ascii file input/output both
in the radmc3dData class and also to the readData and readGrid
functions.

Added a nearest neighbour interpolator for the octrees.

Added functions to find the index of a leaf that contains a specific
coordinate/point.

Added an example decision function to be used in the AMR tree building.