Model validator function + small things
The validateModel() function in the setup module checks the functions
inside a model for consistency.

Coordinates to the AMR cell refinement decision function will be passed
as regular arguments instead of keyword arguments.

The AMR refinement decision function can now also be implemented in a
model. In makeSpatialGrid() the dfunc keyword argument is checked and if
its None then the model is checked if it has a function called
"decisionFunction()" which will then be used instead of dfunc.

Added the grid_style member to the radmc3dGrid and radmc3dOctree
classes.