Image fits output for single frequency + other fix
Added the possibility to write out single frequency frame of a multi
frequency image cube.

Changed the module name of coordinate transformation module back to
crd_trans for consistency

Fixed a bug in the convolution function that caused the wrong pixel
scale to be used for the kernel generation. The conditional was the
opposite I intended to be used so the pixel size was assumed to be in au
when it was in fact arcsec and vice versa.

Fixed the natcont module import in the staratm module.

In the setup module fixed some unicode issues with strings.

Updated the version number to v0.30