Fixed setup.py
At some point I must have changed the setup.py script and totally forgot
about it but it wasn't working properly. There was a bug in the function
that is supposed to automatically find all modules and python scripts
and it always returned an empty list.

Now it is fixed.