Fixed python2 compatibility with setuptools
The installation with setuptools was broken under python2 as for python3
compatibility I converted the packageNames to unicode, which python2
didn't understand. So the final package name list that went to the setup
function had to be converted back to string. Now it should work fine
with both python 2 and python 3