Replaced readline() to fromfile() in ascii readers
The problem with reading formatted ascii files with most methods
(including readline) is that the file structure must be known
beforehand and uniform in the whole file. I.e. It's not straightforward
to read a file with three integers in the first three lines, followed
by four column of floats with 300 lines, followed by a single column of
floats, whih is again followed by a 6 column of floats with 1000 lines.
This is the file structure of the dustkapscatmat_xxxxx.inp files.

Such files can only be read line-by-line, then get the individual
elements / columns by splitting the line by a separator of a blank
space. The problem is if the file contains a couple of blank lines.
It is of course possible to add a conditional to check if the line
is blank, but since one has to check that for every single line
that would slow things down significantly (I guess, I actually never
checked it.. ). So I the simplest and most convenient way to deal
with such files is to use numpy's fromfile() function with a blank space
as separator. This way the whole file will be read line-by-line into
a single numpy array. The nice thing is that fromfile() won't care about
the blank lines or multiple blank spaces, etc. It will get rid of them
automatically. So we can read the whole file at once into a float array,
split off the header and convert it back to integers, then slice up the
remainder of the float array according to the header, which contains the
information on the data size/shape in the file.

I converted the regular grid readers, the dust opacity readers and the
data readers to read the formatted ascii files this way. The only
remaining part to be converted like this is the octree reader.