BUGFIX: gas scalar fields were not read properly
There was a bug in radmc3dData._scalarfieldReader(). This is related to
the changed I've made according to Kees' suggestion to change the way
data files are read. Previously I passed the file name to numpy's
fromfile function and read the header first, then if it was consistent
with the spatial grid, I read the whole file. Of course I had to throw
away the first 3-4 elements of the big array, which was the header.
However, this way I did not need to know how long the header really is
(3 or 4 element) as I could test it.

When I replaced the file name passed to the fromfile function to the
file object I forgot to change the test. So dust data were read just
fine but gas data were not as those have only 3 element headers.

I changed the test and there is now an additional "ndim" argument passed
to the _scalarfieldReader() function to tell how many dimensions the
data I expect to have (4 for dust 3 for gas). It should now work fine.