Bugfix in radmc3dData.getTau()
In the interpolation of the dust opacities it was not checked if the
wavelength is stored in ascending order. If not the interpolator always
took the first element of the interpolated array as the interpolated
value. This is now fixed. If the dust opacity is stored in descending
order (in wavelength!) then the arrays are reversed for the
interpolation.