BUGFIX: Image convolution always returned positive image
In the inverse FFT abs() was used to convert the complex array returned
by the inverse FFT to a real valued (float64) array. This is not a
problem for total intensity images where by definition the intensity
must not be negative. However, it is a problem for Stokes images which
can be negative. abs() has now be replaced with real().